+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Strip all logged lines of extra whitespace
 - Align logged lines with four space indents
 - Changed default axis timestamp labels to be more readable
+- Strip all logged lines of extra whitespace
+- Align logged lines with four space indents
 
 ## 2.0.0 - Breaking changes