+++ CHANGELOG.md
@@ -4,10 +4,9 @@
 
 - Strip all logged lines of extra whitespace
 - Align logged lines with four space indents
-- Changed default axis timestamp labels to be more readable
-- Strip all logged lines of extra whitespace
-- Align logged lines with four space indents
 - Do not automatically create subplots figure and axis in Figure
+- Changed default axis timestamp labels to be more readable
+- Support lists in moving average functions
 
 ## 2.0.0 - Breaking changes